A hacky class combiner type class
The idea is I want two or more objects be able to call methods.

Take whatever data that current object might need from method.

Or as well as allow that method to get data from the object that called it.

This is kinda hard to explain, but if this works like I want it to.

This means easier access between our ConfigControl & Addons.